added custom schemas
