+++ PluginZohoCreator/Helper/Settings.cs
@@ -1,10 +1,12 @@
 using System;
+using System.Collections.Generic;
 
 namespace PluginZohoCreator.Helper
 {
     public class Settings
     {
         public string Token { get; set; }
+        public List<CustomUrlObject> CustomSchemaList { get; set; }
 
         /// <summary>
         /// Validates the settings input object
@@ -17,5 +19,13 @@ namespace PluginZohoCreator.Helper
                 throw new Exception("the Token property must be set");
             }
         }
+    }
+
+    public class CustomUrlObject
+    {
+        public string ApplicationOwner { get; set; }
+        public string ApplicationName { get; set; }
+        public string FormName { get; set; }
+        public string ViewName { get; set; }
     }
 }
